+++ apex-checks/src/main/java/org/fundacionjala/enforce/sonarqube/apex/checks/DmlInConstructorCheck.java
@@ -23,15 +23,12 @@
  */
 package org.fundacionjala.enforce.sonarqube.apex.checks;
 
-import com.sonar.sslr.api.AstNode;
-import com.sonar.sslr.api.Grammar;
 import org.sonar.api.server.rule.RulesDefinition;
 import org.sonar.check.Priority;
 import org.sonar.check.Rule;
 import org.sonar.squidbridge.annotations.ActivatedByDefault;
 import org.sonar.squidbridge.annotations.SqaleConstantRemediation;
 import org.sonar.squidbridge.annotations.SqaleSubCharacteristic;
-import org.sonar.squidbridge.checks.SquidCheck;
 
 import org.fundacionjala.enforce.sonarqube.apex.api.grammar.RuleKey;
 
@@ -48,12 +45,12 @@ import org.fundacionjala.enforce.sonarqube.apex.api.grammar.RuleKey;
 @SqaleSubCharacteristic(RulesDefinition.SubCharacteristics.READABILITY)
 @SqaleConstantRemediation("5min")
 @ActivatedByDefault
-public class DmlInConstructorCheck extends SquidCheck<Grammar> {
+public class DmlInConstructorCheck extends DmlStatementCheck {
 
     /**
      * Stores a message template.
      */
-    public static final String MESSAGE = "The DML statement \"%s\", can not be inside a constructor";
+    private static final String MESSAGE = "The DML statement \"%s\", can not be inside a constructor";
 
     /**
      * It is the code of the rule for the plugin.
@@ -63,22 +60,8 @@ public class DmlInConstructorCheck extends SquidCheck<Grammar> {
     /**
      * The variables are initialized and subscribe the base rule.
      */
-    @Override
-    public void init() {
-        subscribeTo(RuleKey.CONSTRUCTOR_DECLARATION);
-    }
-
-    /**
-     * It is responsible for verifying whether the rule is met in the rule base.
-     * In the event that the rule is not correct, create message error.
-     *
-     * @param astNode It is the node that stores all the rules.
-     */
-    @Override
-    public void visitNode(AstNode astNode) {
-        if (astNode.hasDescendant(RuleKey.DML_STATEMENT)) {
-            getContext().createLineViolation(this, String.format(MESSAGE,
-                    astNode.getFirstDescendant(RuleKey.DML_STATEMENT).getTokenValue()), astNode);
-        }
+    public DmlInConstructorCheck() {
+        ruleKey = RuleKey.CONSTRUCTOR_DECLARATION;
+        message = MESSAGE;
     }
 }
